Update file delimiter overlay requirements
There shouldn't be two ways to specify the same thing.

E.g. leave data_row to default to 1, but if the person specifies data_row=1 this is the same thing. Now you have two ways to specify the same thing - BUT they will have different SAIDs.

We want to try and ensure that something can only be said one way (e.g. if you want data to start on row one, you can only say it data_row=1).

We have gone through these keys and set them to required which forces people to be explicit but the rest of the rules need to be checked first.